requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-mongodb
pymongo
unstructured
python-docx
kafka-python
tqdm
